+++ src/settings/display.ts
@@ -7,12 +7,13 @@ import {moveIndex} from "../utils";
 
 // TODO: settings with a site pattern/regex (powerful and easy to verify - isActive indicator)
 // TODO: escape input examples in description
-const customRulesMap: Map<Target, CustomDisplayRule[]> = new Map([
-    [Target.Scene, [{pattern: "1", filter: "organized:true", display: {color: "purple"}}, {pattern: "2", filter: "organized:true", display: {color: "purple"}}, {pattern: "3", filter: "organized:true", display: {color: "purple"}}, {pattern: "4", filter: "organized:true", display: {color: "purple"}}, {pattern: "5", filter: "file_count:{value:1,modifier:GREATER_THAN}", display: {color: "brown"}}]],
-    [Target.Studio, [{pattern: "*", filter: "scene_count:{value:5,modifier:GREATER_THAN}", display: {color: "purple"}}]]
-]);
+const customRules: CustomDisplayRule[] = [
+    {target: Target.Scene, pattern: "*", filter: "organized:false", display: {color: "blue"}},
+    {target: Target.Scene, pattern: "*://*.javlibrary.com/*", filter: "", display: {color: "purple"}},
+    {target: Target.Studio, pattern: "*://stashdb.org/*", filter: "scene_count:{value:5,modifier:GREATER_THAN}", display: {color: "purple"}}
+];
 
-export const customDisplayRules: Map<Target, CustomDisplayRule[]> = await getValue<Map<Target, CustomDisplayRule[]>>(StorageKey.CustomDisplayRules, customRulesMap);
+export const customDisplayRules: CustomDisplayRule[] = await getValue<CustomDisplayRule[]>(StorageKey.CustomDisplayRules, customRules);
 
 export function initDisplaySettings() {
     let description = "Custom display rules can change the display of check marks. " +
@@ -21,7 +22,8 @@ export function initDisplaySettings() {
         "The order can be changed by dragging. " +
         "If no rule applies, the default display options are use. " +
         "GraphQL filters may not contain AND/OR/NOT. " +
-        "Multiple filters can still be concatenated by ','."
+        "Multiple filters can still be concatenated by ','. " +
+        "Leave the filter empty to always apply."
     let displaySection = newSettingsSection("display", "Display", description);
     populateDisplaySection(displaySection);
 }
@@ -37,29 +39,28 @@ function populateDisplaySection(displaySection: HTMLElement) {
     let tableBody = document.createElement("tbody");
     table.append(tableBody);
     let tableHeading = document.createElement("h2");
-    tableHeading.innerHTML = readable(Target.Scene)
+    tableHeading.innerHTML = "Custom Display Rules";
     displaySection.append(tableHeading);
     displaySection.append(table);
     Sortable.create(tableBody, {
         onEnd: event => {
             if (event.oldIndex && event.newIndex) {
-                let old = customDisplayRules.get(Target.Scene) ?? [];
-                customDisplayRules.set(Target.Scene, moveIndex(old, event.oldIndex, event.newIndex));
+                moveIndex(customDisplayRules, event.oldIndex, event.newIndex);
             }
         }
     });
-    populateCustomRulesTable(tableBody, Target.Scene);
+    populateCustomRulesTable(tableBody);
 }
 
-function populateCustomRulesTable(tableBody: HTMLTableSectionElement, target: Target) {
-    let tableRows = Array.from(customDisplayRules.get(target) ?? [])
+function populateCustomRulesTable(tableBody: HTMLTableSectionElement) {
+    let tableRows = Array.from(customDisplayRules)
         .map(tableRow);
     tableBody.replaceChildren(...tableRows);
 }
 
 function tableHeadRow(): HTMLTableRowElement {
     let row = document.createElement("tr");
-    row.innerHTML = "<th>URL Pattern</th><th>GraphQL Filter</th><th>Color</th><th>Preview</th>";
+    row.innerHTML = "<th>Type</th><th>URL Pattern</th><th>GraphQL Filter</th><th>Color</th><th>Preview</th>";
     return row;
 }
 
@@ -74,6 +75,7 @@ function tableRow(customRule: CustomDisplayRule): HTMLTableRowElement {
     previewElement.append(preview)
 
     row.append(
+        dataCell(customRule.target),
         dataCell(customRule.pattern),
         dataCell(customRule.filter),
         dataCell(customRule.display.color),