+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 
-## [1.1.0]
+## [WIP]
 
-- Add support for websites: manyvids, warashi-asian-pornstars.fr
+- Add support for websites: manyvids, warashi-asian-pornstars.fr, hobby.porn, pornhub
 - Add showCheckMark setting
 - Add organized flag to tooltip
 - Use tippy.js for better tooltips